pullRequestNumber Long to Int

+++ selfouettie-server/src/main/scala/net/yoshinorin/selfouettie/models/db/PullRequestReviewCommentEvents.scala
@@ -6,7 +6,7 @@ case class PullRequestReviewCommentEvents(
   eventId: Long,
   userName: String,
   repositoryId: Long,
-  pullRequestNumber: Long,
+  pullRequestNumber: Int,
   action: String,
   createdAt: Long
 ) extends BaseEvent[PullRequestReviewCommentEvents]
